refactor : 주문 상태를 전략 패턴을 통해 동적으로 주입받아 관리하도록 변경
