--- src/main/java/com/yes25/yes255orderpaymentserver/application/service/OrderStatusService.java
@@ -1,9 +1,0 @@
-package com.yes25.yes255orderpaymentserver.application.service;
-
-import com.yes25.yes255orderpaymentserver.presentation.dto.request.UpdateOrderRequest;
-import com.yes25.yes255orderpaymentserver.presentation.dto.response.UpdateOrderResponse;
-
-public interface OrderStatusService {
-
-    UpdateOrderResponse updateOrderStatusByOrderId(String orderId, UpdateOrderRequest request, Long userId);
-}
