+++ src/main/java/com/yes25/yes255orderpaymentserver/application/service/strategy/payment/impl/KakaoPayment.java
@@ -1,13 +1,12 @@
-package com.yes25.yes255orderpaymentserver.application.service.strategy.impl;
+package com.yes25.yes255orderpaymentserver.application.service.strategy.payment.impl;
 
-import com.yes25.yes255orderpaymentserver.application.service.strategy.PaymentStrategy;
+import com.yes25.yes255orderpaymentserver.application.service.strategy.payment.PaymentStrategy;
 import com.yes25.yes255orderpaymentserver.presentation.dto.request.CreatePaymentRequest;
 import com.yes25.yes255orderpaymentserver.presentation.dto.response.CreatePaymentResponse;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 
-@Component("naver")
-public class NaverPayment implements PaymentStrategy {
+@Component("kakao")
+public class KakaoPayment implements PaymentStrategy {
 
     @Override
     public CreatePaymentResponse createPayment(CreatePaymentRequest request) {
