+++ src/main/java/com/yes25/yes255orderpaymentserver/presentation/controller/OrderController.java
@@ -2,8 +2,8 @@ package com.yes25.yes255orderpaymentserver.presentation.controller;
 
 import com.yes25.yes255orderpaymentserver.application.dto.response.ReadPurePriceResponse;
 import com.yes25.yes255orderpaymentserver.application.service.OrderService;
-import com.yes25.yes255orderpaymentserver.application.service.OrderStatusService;
 import com.yes25.yes255orderpaymentserver.application.service.PreOrderService;
+import com.yes25.yes255orderpaymentserver.application.service.context.OrderStatusContext;
 import com.yes25.yes255orderpaymentserver.common.jwt.HeaderUtils;
 import com.yes25.yes255orderpaymentserver.common.jwt.JwtUserDetails;
 import com.yes25.yes255orderpaymentserver.common.jwt.annotation.CurrentUser;
@@ -44,8 +44,8 @@ public class OrderController {
 
     private final PreOrderService preOrderService;
     private final OrderService orderService;
-    private final OrderStatusService orderStatusService;
-
+    private final OrderStatusContext orderStatusContext; 
+    
     @Operation(summary = "주문 생성", description = "새로운 주문을 생성합니다.")
     @ApiResponses(value = {
         @ApiResponse(responseCode = "200", description = "주문 생성 성공", content = @Content(schema = @Schema(implementation = CreateOrderResponse.class))),
@@ -121,7 +121,7 @@ public class OrderController {
     public ResponseEntity<UpdateOrderResponse> update(@PathVariable String orderId, @RequestBody UpdateOrderRequest request, @CurrentUser JwtUserDetails jwtUserDetails) {
         return ResponseEntity.ok()
             .headers(HeaderUtils.addAuthHeaders(jwtUserDetails))
-            .body(orderStatusService.updateOrderStatusByOrderId(orderId, request, jwtUserDetails.userId()));
+            .body(orderStatusContext.updateOrderStatusByOrderId(orderId, request, jwtUserDetails.userId()));
     }
 
     @Operation(summary = "주문 배송 정보 조회", description = "주문 ID로 배송 정보를 조회합니다.")
